Move codemonitoring resolvers in cmd/frontend
Not needed by other services, so it should live here.
